Перетянуты изменения из версии 36.050. В проводном режиме проверено следующее: 1. Сброс в заводские уставки 2. Редактирование параметров. 3. Работа с уровнями доступа. 4. Выбор параметров для статусного меню. 5. Задание частоты. 6. Изменение языка. 7. НЕ работают команды ПУСК и СТОП.

+++ P24/HrdDrivers/uart.h
@@ -104,7 +104,7 @@ void rt_hw_usart_init(void);
 inline void startRxIdle(portUart_type *uart);
 inline void startRxReceive(portUart_type *uart);
 
-inline void TIMuartWdStart( u16 );
+inline void TIMuartWdStart( u32 );
 inline void TIMuartWdSop(void);
 
 #endif // UART_H
